Update asuz backups to run as root
This allows them to run from cron.daily so they catch up when the laptop
wakes up.

I originally wanted backup scripts to run as users, since it's generally
more secure not to run as root. However, in the case of backups I
realized you actually *don't* want users to have access, so that rogue
software running as the user can't mess up the backups. When backups run
as root, we have to trust borg, but we don't have to trust everything
the user runs.