Remove borg backup for photos
This relies on backing up to a local borg instance and then syncing to
Wasabi. That is unsafe because malware could corrupt my local repo and
then it would sync the corruption to Wasabi.

My new photo backup uses append-only duplicity, and my borg backups use
BorgBase so the repo is not on site.