Use public hledger docker image
It's basically the same as what I'm using, hence very few changes. It
does require that the /data directory be writeable, so it can save some
crypto file.

The upside is that this image has an updated version of hledger that
exposes the API, so I can use it to build a custom UI.