Add tls redirect to http services
The weird thing abou this is that it doesn't work *unless* you create a
second router for https. If the same router is serving http and https,
you just get a 404 at the http scheme. This makes the config kind of
ugly, but I guess if it works it works.